app009_simple_image_viewer.py
import streamlit as st
from PIL import Image

st.title("Simple Image Viewer")
uploaded_image = st.file_uploader("Upload an image", type=["png", "jpg", "jpeg"])
if uploaded_image:
    image = Image.open(uploaded_image)
    st.image(image, caption="Uploaded Image", use_column_width=True)
